+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/jira/home/SharedHomeFormula.kt
@@ -2,6 +2,7 @@ package com.atlassian.performance.tools.awsinfrastructure.jira.home
 
 import com.atlassian.performance.tools.aws.api.Storage
 import com.atlassian.performance.tools.awsinfrastructure.AwsCli
+import com.atlassian.performance.tools.awsinfrastructure.api.hardware.Computer
 import com.atlassian.performance.tools.infrastructure.api.jira.JiraHomeSource
 import com.atlassian.performance.tools.infrastructure.api.jira.SharedHome
 import com.atlassian.performance.tools.infrastructure.api.os.Ubuntu
@@ -12,15 +13,17 @@ internal class SharedHomeFormula(
     private val pluginsTransport: Storage,
     private val jiraHomeSource: JiraHomeSource,
     private val ip: String,
-    private val ssh: Ssh
+    private val ssh: Ssh,
+    private val computer: Computer
 ) {
     private val localSubnet = "10.0.0.0/24"
-    private val localSharedHome = "/opt/jira-shared-home"
+    private val localSharedHome = "/home/ubuntu/jira-shared-home"
 
     private val ubuntu = Ubuntu()
 
     fun provision(): SharedHome {
         ssh.newConnection().use {
+            computer.setUp(it)
             val jiraHome = jiraHomeSource.download(it)
             AwsCli().download(
                 location = pluginsTransport.location,